fix: DeltaFreqHz int type consistency; Message column 40 chars; CORS from config; TrayOrchestrator dispose wait
- WsjtxDecodeDto.DeltaFreqHz changed from uint to int to align with PostDecodeDto
- Added (int) casts in Worker.cs and TrayOrchestrator.cs where ParsedDecode.DeltaFreqHz (uint) is assigned
- WsjtxDecodeConfiguration Message column increased from 30 to 40 chars; migration IncreaseWsjtxMessageLength generated and applied
- CORS switched from hardcoded named policy to AddDefaultPolicy reading Cors:AllowedOrigins from config; appsettings.json updated
- TrayOrchestrator.Dispose now waits up to 3 s for _runTask before disposing _httpClient to prevent ObjectDisposedException

+++ backend/HamHub.WsjtxService/Worker.cs
@@ -95,7 +95,7 @@ public class Worker : BackgroundService
                 DxGrid: dxGrid,
                 Snr: decode.Snr,
                 DeltaTime: decode.DeltaTime,
-                DeltaFreqHz: decode.DeltaFreqHz,
+                DeltaFreqHz: (int)decode.DeltaFreqHz,
                 FrequencyMhz: freqMhz,
                 Mode: decode.Mode,
                 DecodedAt: DateTime.UtcNow
